+++ tsconfig.json
@@ -1,0 +1,20 @@
+{
+    "compilerOptions": {
+      "target": "ES2020",
+      "module": "commonjs",
+      "moduleResolution": "node",
+      "outDir": "dist",
+      "declaration": false,
+      "removeComments": false,
+      "esModuleInterop": true,
+      "experimentalDecorators": true,
+      "skipLibCheck": true,
+    },
+    "include": [
+      "src",
+    ],
+    "exclude": [
+      "node_modules",
+      "**/__tests__/*"
+    ]
+  }
